src/app/requirements.txt
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-WTF==1.0.1
bootstrap-flask==2.2.0
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
google-cloud-aiplatform == 1.28.0
google-cloud-pubsub == 2.18.0
futures==3.0.5
vertexai==0.0.1
